Move proto functional freceiver to experimental.hs
I wrote some prototype code for a frame receiver that is written in a
more functional style. Hopefully the plan from here is to get the
current codebase up to scratch with h2spec, then see if we can migrate
the control flow related to frame processing to be a written in a bit
more evaluational style.